+++ server/src/main/java/org/elasticsearch/node/ResponseCollectorService.java
@@ -152,7 +152,7 @@ public final class ResponseCollectorService implements ClusterStateListener {
             // the concurrency compensation is defined as the number of
             // outstanding requests from the client to the node times the number
             // of clients in the system
-            double concurrencyCompensation = (double) outstandingRequests * clientNum;
+            double concurrencyCompensation = outstandingRequests * clientNum;
 
             // Cubic queue adjustment factor. The paper chose 3 though we could
             // potentially make this configurable if desired.
