+++ server/src/main/java/org/elasticsearch/search/aggregations/bucket/composite/DoubleValuesSource.java
@@ -44,7 +44,7 @@ class DoubleValuesSource extends SingleDimensionValuesSource<Double> {
 
     @Override
     void copyCurrent(int slot) {
-        values = bigArrays.grow(values, (long) slot+1);
+        values = bigArrays.grow(values, slot+1);
         if (missingBucket && missingCurrentValue) {
             bits.clear(slot);
         } else {
